Renames parametric to non_parametric, changes relative to absolute imports, changes location of pipeline template from core to pipeline
