--- pandora/core/model/__init__.py
@@ -1,8 +1,0 @@
-from .builder import ParametricModelBuilder
-from .base import ModelBuilder
-
-
-__all__ = [
-    'ModelBuilder',
-    'ParametricModelBuilder',
-]
