+++ pandora/core/model/builder/__init__.py
@@ -1,0 +1,8 @@
+from .non_parametric import NonParametricModelBuilder
+from .base import ModelBuilder
+
+
+__all__ = [
+    'ModelBuilder',
+    'NonParametricModelBuilder',
+]
