+++ pandora/core/model/estimators/__init__.py
@@ -1,0 +1,9 @@
+from xgboost import XGBRegressor, XGBClassifier
+
+
+ESTIMATOR_ALIAS = {
+    'xgbr': XGBRegressor,
+    'xgbc': XGBClassifier
+}
+
+__all__ = ['ESTIMATOR_ALIAS']
