+++ pandora/pipeline/template/base.py
