+++ pandora/pipeline/template/image.py
@@ -1,4 +1,4 @@
-from .base import Template
+from pandora.pipeline.template.base import Template
 
 
 class ImageTemplate(Template):
